# openapi.yaml
openapi: 3.0.3
info:
    title: MechHub Miniback API
    version: 1.0.0
    description: Standalone backend for MechHub. Authentication is session-based and uses the `mechhub.sid` cookie.
servers:
    - url: http://localhost:3001
      description: Local development server
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: mechhub.sid
    schemas:
        User:
            type: object
            required:
                - id
                - email
                - displayName
                - avatarUrl
                - bio
                - defaultRole
                - createdAt
            properties:
                id:
                    type: integer
                    example: 1
                email:
                    type: string
                    format: email
                    example: test@example.com
                displayName:
                    type: string
                    example: MechFox4821
                avatarUrl:
                    type: string
                    example: https://api.dicebear.com/9.x/pixel-art/svg?seed=MechFox4821
                bio:
                    type: string
                    example: Building the next mechanics education workflow.
                defaultRole:
                    type: string
                    enum: [teacher, student]
                    example: student
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:00:00.100Z
...
        ErrorResponse:
            type: object
            required:
                - message
            properties:
                message:
                    type: string
                    example: Not authenticated
        Class:
            type: object
            required:
                - id
                - name
                - description
                - ownerUserId
                - inviteCode
                - status
                - createdAt
                - membershipRole
                - isOwner
            properties:
                id:
                    type: integer
                    example: 1
                name:
                    type: string
                    example: Statics 101
                description:
                    type: string
                    example: Introductory statics class for spring term.
                ownerUserId:
                    type: integer
                    example: 1
                inviteCode:
                    type: string
                    example: 8F1A2C4D
                status:
                    type: string
                    enum: [active, archived]
                    example: active
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:10:00.200Z
                membershipRole:
                    type: string
                    enum: [teacher, student]
                    nullable: true
                    example: teacher
                isOwner:
                    type: boolean
                    example: true
...
        ClassMember:
            type: object
            required:
                - id
                - classId
                - userId
                - role
                - joinedAt
                - isOwner
                - user
            properties:
                id:
                    type: integer
                    example: 1
                classId:
                    type: integer
                    example: 1
                userId:
                    type: integer
                    example: 1
                role:
                    type: string
                    enum: [teacher, student]
                    example: teacher
                joinedAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:10:00.250Z
                isOwner:
                    type: boolean
                    example: true
                user:
                    $ref: "#/components/schemas/User"
...
        TeacherClassDashboardItem:
            type: object
            required:
                - class
                - studentCount
                - assignmentCount
                - pendingReviewCount
                - upcomingDueCount
            properties:
                class:
                    $ref: "#/components/schemas/Class"
                studentCount:
                    type: integer
                    example: 32
                assignmentCount:
                    type: integer
                    example: 8
                pendingReviewCount:
                    type: integer
                    example: 14
                upcomingDueCount:
                    type: integer
                    example: 2
        TeacherClassesDashboard:
            type: object
            required:
                - summary
                - classes
            properties:
                summary:
                    type: object
                    required:
                        - classCount
                        - studentCount
                        - assignmentCount
                        - pendingReviewCount
                    properties:
                        classCount:
                            type: integer
                            example: 3
                        studentCount:
                            type: integer
                            example: 96
                        assignmentCount:
                            type: integer
                            example: 24
                        pendingReviewCount:
                            type: integer
                            example: 31
                classes:
                    type: array
                    items:
                        $ref: "#/components/schemas/TeacherClassDashboardItem"
        TeacherAssignmentDashboardItem:
            type: object
            required:
                - assignmentId
                - classId
                - className
                - title
                - status
                - dueAt
                - totalStudents
                - submittedCount
                - reviewedCount
                - pendingReviewCount
            properties:
                assignmentId:
                    type: integer
                    example: 9
                classId:
                    type: integer
                    example: 1
                className:
                    type: string
                    example: Statics 101
                title:
                    type: string
                    example: Week 3 Force Balance Homework
                status:
                    type: string
                    enum: [draft, published, closed, archived]
                    example: published
                dueAt:
                    type: string
                    format: date-time
                    nullable: true
                    example: 2026-03-30T15:59:59.000Z
                totalStudents:
                    type: integer
                    example: 32
                submittedCount:
                    type: integer
                    example: 28
                reviewedCount:
                    type: integer
                    example: 17
                pendingReviewCount:
                    type: integer
                    example: 11
        TeacherAssignmentsDashboard:
            type: object
            required:
                - summary
                - assignments
            properties:
                summary:
                    type: object
                    required:
                        - assignmentCount
                        - publishedCount
                        - pendingReviewCount
                        - overdueCount
                    properties:
                        assignmentCount:
                            type: integer
                            example: 24
                        publishedCount:
                            type: integer
                            example: 18
                        pendingReviewCount:
                            type: integer
                            example: 31
                        overdueCount:
                            type: integer
                            example: 4
                assignments:
                    type: array
                    items:
                        $ref: "#/components/schemas/TeacherAssignmentDashboardItem"
        ClassDashboard:
            type: object
            required:
                - class
                - summary
                - members
                - assignments
            properties:
                class:
                    $ref: "#/components/schemas/Class"
                summary:
                    type: object
                    required:
                        - studentCount
                        - assignmentCount
                        - pendingReviewCount
                    properties:
                        studentCount:
                            type: integer
                            example: 32
                        assignmentCount:
                            type: integer
                            example: 8
                        pendingReviewCount:
                            type: integer
                            example: 14
                members:
                    type: array
                    items:
                        $ref: "#/components/schemas/ClassMember"
                assignments:
                    type: array
                    items:
                        $ref: "#/components/schemas/TeacherAssignmentDashboardItem"
        AssignmentClassDashboard:
            type: object
            required:
                - assignment
                - summary
                - submissions
            properties:
                assignment:
                    $ref: "#/components/schemas/Assignment"
                summary:
                    type: object
                    required:
                        - studentCount
                        - submittedCount
                        - reviewedCount
                        - pendingReviewCount
                        - lateCount
                    properties:
                        studentCount:
                            type: integer
                            example: 32
                        submittedCount:
                            type: integer
                            example: 28
                        reviewedCount:
                            type: integer
                            example: 17
                        pendingReviewCount:
                            type: integer
                            example: 11
                        lateCount:
                            type: integer
                            example: 3
                submissions:
                    type: array
                    items:
                        $ref: "#/components/schemas/AssignmentSubmission"
...
        StudentAssignmentsDashboard:
            type: object
            required:
                - summary
                - assignments
            properties:
                summary:
                    type: object
                    required:
                        - pendingCount
                        - dueSoonCount
                        - submittedCount
                        - reviewedCount
                    properties:
                        pendingCount:
                            type: integer
                            example: 3
                        dueSoonCount:
                            type: integer
                            example: 1
                        submittedCount:
                            type: integer
                            example: 6
                        reviewedCount:
                            type: integer
                            example: 4
                assignments:
                    type: array
                    items:
                        $ref: "#/components/schemas/StudentAssignmentCard"
...
paths:
...
    /dashboards/teacher/classes:
        get:
            tags:
                - Dashboards
            summary: Get the teacher's class overview dashboard
            operationId: getTeacherClassesDashboard
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Teacher class dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TeacherClassesDashboard"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only teachers can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /dashboards/teacher/assignments:
        get:
            tags:
                - Dashboards
            summary: Get the teacher's assignment overview dashboard
            operationId: getTeacherAssignmentsDashboard
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Teacher assignment dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TeacherAssignmentsDashboard"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only teachers can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /dashboards/classes/{classId}:
        get:
            tags:
                - Dashboards
            summary: Get a teacher-facing class dashboard
            operationId: getClassDashboard
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Class dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ClassDashboard"
                "400":
                    description: Invalid class id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only class teachers can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /dashboards/classes/{classId}/assignments/{assignmentId}:
        get:
            tags:
                - Dashboards
            summary: Get a teacher-facing assignment dashboard for one class
            operationId: getClassAssignmentDashboard
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
                - in: path
                  name: assignmentId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Assignment dashboard for a class
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AssignmentClassDashboard"
                "400":
                    description: Invalid class id or assignment id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only class teachers can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class or assignment not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /dashboards/student/assignments:
        get:
            tags:
                - Dashboards
            summary: Get the current student's assignment dashboard
            operationId: getStudentAssignmentsDashboard
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Student assignments dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/StudentAssignmentsDashboard"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only students can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...